Fix jitter backoff to always use uniform interval
Jitter backoff now increases the backoff using a similar algorithm to
normal, except that the new delay is chosen uniformly from an interval
around the increment:

[0.5 * backoff:increment(N), 1.5 * backoff:increment(N)]

When a ceiling value (Max) is used and the delay is above Max div 3 the
following interval is used: [Max div 3, Max]. This means that once a
delay is above Max div 3 all subsequent delays will be from that
interval.

The values 0.5 and 1.5 are suggested in:
The Synchronization of Periodic Routing Messages,
Sally Floyd and Van Jacobson, 1994, http://ee.lbl.gov/papers/sync_94.pdf

Jitter nolonger uses application env `rand_factor` and `def_multiplier`.